# docs/openapi/openapi.v0.12.0.yaml
openapi: 3.0.3
info:
  title: SaaS Boilerplate API
  version: 0.12.0
  description: Organisation-first SaaS boilerplate API. v0.12.0 completes the OIDC
    / SSO build by activating authorization-code + PKCE start/callback/logout routes,
    OIDC discovery, token exchange, RS256 id_token validation, auto-provisioning,
    group-to-role mapping and audit/security events for SSO login.
servers:
- url: http://localhost:4000/api
  description: Local development API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Generic bearer auth scheme.
    UserSessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT user_session_token
      description: Short-lived logged-in user token used by the frontend and Postman
        during development.
    PersonalAccessTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT personal_access_token
      description: Long-lived user-owned scoped token for Postman, scripts and user
        automation.
    OrganisationApiTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT organisation_api_token
      description: Organisation-owned token for external/system-to-system integrations.
        Uses external audience and external:* scopes.
    ServiceAccountTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: Service Account Token
      description: Reserved for later service identity requirements.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            details:
              type: object
...
paths:
...
  /org/sso-config/test:
    post:
      tags:
      - SSO
      summary: Validate organisation SSO config
      description: 'Runs local validation against saved SSO provider configuration.
        Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope:
        organisation.sso.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Validate organisation SSO config succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Validate organisation SSO config created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /platform/plans:
    get:
      tags:
      - Internal
      summary: List plan catalogue
      description: 'Returns SaaS plan catalogue, feature defaults and limits. No payment
        gateway is active in v0.9.1. Authentication: UserSessionAuth or PersonalAccessTokenAuth.
        Required permission/scope: plans.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      responses:
        '200':
          description: List plan catalogue succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Internal
      summary: Create or upsert plan
      description: 'Creates/updates a plan catalogue item for subscription readiness.
        Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope:
        plans.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Create or upsert plan succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Create or upsert plan created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /platform/plans/{planKey}:
    get:
      tags:
      - Internal
      summary: Get plan by key
      description: 'Returns one plan catalogue item. Authentication: UserSessionAuth
        or PersonalAccessTokenAuth. Required permission/scope: plans.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      responses:
        '200':
          description: Get plan by key succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Internal
      summary: Update plan by key
      description: 'Updates plan pricing placeholders, feature defaults or limits.
        Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope:
        plans.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Update plan by key succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Update plan by key created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /org/plan:
    get:
      tags:
      - Internal
      summary: Get current organisation plan
      description: 'Returns the authenticated organisation''s plan assignment, effective
        limits and feature readiness metadata. Authentication: UserSessionAuth or
        PersonalAccessTokenAuth. Required permission/scope: plans.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      responses:
        '200':
          description: Get current organisation plan succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /platform/organisations/{organisationId}/plan:
    get:
      tags:
      - Internal
      summary: Get platform organisation plan
      description: 'Returns plan assignment for a requested organisation. Authentication:
        UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope: plans.view.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      responses:
        '200':
          description: Get platform organisation plan succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - Internal
      summary: Assign organisation plan
      description: 'Assigns a plan and optionally applies plan feature defaults and
        limits. Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required
        permission/scope: plans.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Assign organisation plan succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Assign organisation plan created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /platform/organisations/{organisationId}/plan/apply-defaults:
    post:
      tags:
      - Internal
      summary: Apply organisation plan defaults
      description: 'Applies assigned plan feature defaults and limits to the organisation.
        Authentication: UserSessionAuth or PersonalAccessTokenAuth. Required permission/scope:
        plans.manage.'
      security:
      - UserSessionAuth: []
      - PersonalAccessTokenAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Apply organisation plan defaults succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '201':
          description: Apply organisation plan defaults created resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  data:
                    type: object
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission or scope error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...